fix bugs

+++ piton/utils/installer.py
@@ -25,11 +25,9 @@ def remove(package):
 	pending_removals.append(metadata.dist_info)
 	for pending_removal in pending_removals:
 		try:
-			shutil.rmtree(os.path.join("python_modules", pending_removal))
+			path = os.path.join("python_modules", pending_removal)
+			if os.path.isdir(path):
+				shutil.rmtree(os.path.join("python_modules", pending_removal))
 		except:
-			# try:
-			# 	os.remove(os.path.join("python_modules", pending_removal+".py"))
-			# except:
-			# 	pass
 			raise("cannot remove")
 	return True
